reslove comments
