+++ mmpose/models/heads/__init__.py
@@ -1,9 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .base_head import BaseHead
-from .heatmap_heads import (HeatmapHead, MultiStageHeatmapHead,
-                            MultiStageMultiUnitHeatmapHead)
+from .heatmap_heads import CPMHead, HeatmapHead, MSPNHead, ViPNASHead
 
-__all__ = [
-    'BaseHead', 'HeatmapHead', 'MultiStageHeatmapHead',
-    'MultiStageMultiUnitHeatmapHead'
-]
+__all__ = ['BaseHead', 'HeatmapHead', 'CPMHead', 'MSPNHead', 'ViPNASHead']
